feat: 옷 <-> 파일 서버 연동 구현중

+++ backend/clothes/src/main/java/com/ssafy/ozz/clothes/global/fegin/file/dto/FeignFileInfo.java
@@ -1,7 +1,7 @@
 package com.ssafy.ozz.clothes.global.fegin.file.dto;
 
 public record FeignFileInfo(
-        int fileId,
+        Long fileId,
         String fileDir,
         String fileName,
         String fileType
